requirements.txt
aio-pika==9.3.1
aiogram==3.2.0
python-dotenv==1.0.0
